+++ src/main/java/plus/dragons/createenchantmentindustry/foundation/mixin/FluidTankBlockMixin.java
@@ -44,7 +44,7 @@ public abstract class FluidTankBlockMixin extends Block implements IBE<BasinBloc
             if (maxSize == 1) {
                 expFluid.drop(serverLevel, VecHelper.getCenterOf(pos), fluidStackBackup.getAmount());
             } else {
-                var total = maxSize * (FluidTankBlockEntity.getCapacityMultiplier() - 1);
+                var total = (maxSize - 1) * FluidTankBlockEntity.getCapacityMultiplier();
                 var leftover = fluidStackBackup.getAmount() - total;
                 if (leftover > 0) {
                     expFluid.drop(serverLevel, VecHelper.getCenterOf(pos), leftover);